+++ MoreFertilizers/MoreFertilizers/docs/Changelog.md
@@ -3,6 +3,10 @@
 #### Todo:
 * Make organic seeds put down organic fertilizer?
 
+#### Version 0.1.6
+* Organic seeds now apply the organic fertilizer.
+* Fixed deshuffling code. Again.
+
 #### Version 0.1.5
 * Joja crops no longer can become organic.
 * Mill preserves organic